chore(log): extract new bug button component
